Reads in encrypted data.
This commit adds the old csspw.py program to make use of some
functions therein. They will soon be merged into the Model and
generalized (e.g. use of GPG instead of bcrypt).

Controller reads now a file 'passwd.bfe' in initialize() that has
to be in the same directory with trove.py (file name and location
will change in the future). It asks for a master passphrase as needed
by bcrypt to decrypt the file and encrypt it again.