Adds search command to controller (and corresponding functions to model and view).
One can now search the list of entries decrypted by the program.
Still missing: if there is only one result after the search, one
should not be prompted to with a selection of one item, but it
should be shown immediately.
Further: when there is no help text available, the password should
be displayed at once.