Updates help text.
We will not use the SHA1 any more to identify the entry to edit/delete.
A standard search and selection menu will be used for that.
However, SHA1 will be used internally to make sure no one tampers with
the text file (this has not been done before in csspw). One could print
a warning if a calculated SHA1 does not match the stored value. This
might show the user that someone has worked on the decrypted file manually
(i.e. without trove).