Changes internal data structure and drops csspw dependency.
This commit changes from a list to a dictionary as internal data structure to
store entries. The keys for the dictionary are the SHA1 hashes of the entry
objects. All remaining functions of csspw (that had been added as a module)
have been incorporated into the Model. The class that represents an entry is
called TroveEntry now instead of ListEntry.