Removes comments with questions in View and Controller.
The argument in all methods starting with do_ is necessary even
if not used in the method. Otherwise the cmd.py module raises a
TypeError. Somehow the argument is connected with what is written
after the do_ (the actual command available to the user). Weird.

In all such methods the additional argument has now the name 'arg'.
All corresponding questions in the comments by Paul were removed.