[controller] defining encrypted file path in constructor
This isn't really the right place for it either, but this allows us to more
easily change the name of the encrypted file such that read_db_file() can be
tested automatically.  At some stage we want to push information such as
this out into a configuration object of some kind.