Warns if only General section exists in trove.conf
The user should in this case create a new encrypted file with a
'create' command which still has to be implemented.
However right now the program still continues to read in the
passwd.bfe file (which has to be in the same directory).
Will have to stop that from happening.