Setting default encrypted_file value before reading config file
This fixes a bug that the value read from the config file always got
overridden by the "passwd.bfe" setting.  Setting the default value earlier
and then overriding it is the right thing to do.  However, the way I've done
this here probably isn't the best way to achieve this