Custom implementation of ConfigParser.
I wanted to have colons as default separators in the config file and
in the encrypted password file. The default of ConfigParser in
Python 2.x is the '=' sign. This commit changes that by subclassing
ConfigParser.py and overriding the 'write' method therein.

The encrypted file is still written 'by hand'. Will see if that can be
changed and still write comments (header) and start and end markers.

Source:
http://stackoverflow.com/questions/4058400/using-colons-in-configparser-python