Removes get_config() method in Model.
Contoller methods now check if a config file is present or not.
The Model however has self.config, an MyConfigParser object.
I was not happy with the way ConfigParser writes the config files
('=' as a delimiter between keys and values, I wanted ':' instead).
MyConfigParser is derived from ConfigParser.ConfigParser() and
overrides the write method.